+++ src/modules/Fixes/FixHeadHitting.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "module/Module.hpp"
+#include <Geode/modify/LevelEditorLayer.hpp>
+
+class $editorModule(FixHeadHitting) {
+    bool onToggled(bool state) override;
+};
+
+class $modify(FHHLevelEditorLayer, LevelEditorLayer) {
+    $registerEditorHooks(FixHeadHitting)
+
+    void onPlaytest();
+};